django_project/requirementsMac.txt
Cheetah==2.4.4
Django==1.10.2
Pillow==3.4.2
PyYAML==3.10
South==0.7.5
Twisted==16.6.0
Unidecode==0.04.19
argparse==1.2.1
boto3==1.4.3
botocore==1.4.93
chardet==2.3.0
colorama==0.2.5
configobj==4.7.2
django-appconf==1.0.2
django-compressor==2.1
django-email-extras==0.3.3
django-filer==1.2.5
django-forms-builder==0.13.0
django-mptt==0.8.6
django-polymorphic==1.0.2
django-registration==2.2
django-storages==1.5.1
docutils==0.13.1
easy-thumbnails==2.3
future==0.15.0
futures==3.0.5
greenlet==0.4.2
gunicorn==17.5
html5lib==0.999
jmespath==0.9.0
jsonpatch==1.3
jsonpointer==1.0
oauth==1.0.1
prettytable==0.7.2
psycopg2==2.4.5
pyserial==2.6
python-dateutil==2.6.0
python-gnupg==0.3.9
rcssmin==1.0.6
requests==2.2.1
rjsmin==1.0.12
s3transfer==0.1.10
six==1.5.2
sphinx-me==0.3
ssh-import-id==3.21
urllib3==1.7.1
virtualenv==1.11.4
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
